comm

+++ database/migrations/2021_03_04_021750_create_products_table.php
@@ -17,11 +17,11 @@ class CreateProductsTable extends Migration
           $table->increments('id');
           $table->string('barcode',50);
           $table->string('name',150);
-          $table->float('buy_price',100);
-          $table->float('sell_price',100);          
+          $table->double('buy_price',100);
+          $table->double('sell_price',100);          
           $table->integer('qty')->default('0');
-          $table->float('total_buy_price',100);
-          $table->float('total_sell_price',100);
+          $table->double('total_buy_price',100);
+          $table->double('total_sell_price',100);
         
           $table->integer('category_id')->unsigned();
           $table->foreign('category_id')->references('id')->on('categories');
